Sign_up test cases refactor
